Refactor STT tests and integrate WAV file loader
- Moved WAV file loader implementation to a dedicated module for better organization.
- Updated end-to-end WAV test to utilize the new WAV file loader.
- Refactored test infrastructure initialization into a separate utility module.
- Replaced mock text injector with real injection processor in the WAV pipeline test.
- Improved error handling and logging in the WAV streaming process.
- Updated WER calculation utilities for better performance and readability.
- Adjusted audio capture thread to use Arc<Mutex> for thread safety.
- Enhanced telemetry metrics to include total request counts.
- Cleaned up unused code and comments across various test files.
- Added detailed documentation for the refactored components.
