+++ crates/app/src/probes/mic_capture.rs
@@ -23,6 +23,7 @@ impl MicCaptureCheck {
         // Prepare ring buffer and spawn capture thread
         let rb = AudioRingBuffer::new(16_384);
         let (audio_producer, audio_consumer) = rb.split();
+        let audio_producer = Arc::new(parking_lot::Mutex::new(audio_producer));
         let (capture_thread, dev_cfg, _config_rx, _device_event_rx) =
             AudioCaptureThread::spawn(config, audio_producer, device_name).map_err(|e| {
                 TestError {
