+++ crates/app/srcs/stt/tests/end_to_end_wav.rs
@@ -1,0 +1,1 @@
+use crate::test_utils::init_test_infrastructure;
